Initial implementation of versioning support
This adds a number of features. `Dispatch#hook()` now accepts four
parameters: `name`, `version`, `options`, and `callback`. The `options`
parameter is an object with two optional properties: `type`, which
allows intercepting only real packets, only fake ones, or all; and
`order`, which is a number where negative values cause the hook to run
earlier and positive values cause it to run later compared to other
hooks.

`Dispatch#unhook()` is currently not implemented.