Remove error-stack-parser (closes #3)
Due to stacktracejs/error-stack-parser#35 parses are incorrect for any
paths which contain a space. Since this project is primarily being run
on Windows machines, this is fairly common case to encounter.

By rolling our own function to parse stack traces, we get rid of the
external dependencies and all the other unnecessary features.