Respect ordering across global and code hooks
This also gets rid of two levels of loops. However, we use a generator
to avoid making a copy of the hook ordering so that we don't keep a
copy of a hook that may be unhooked before it's reached in the future.
Generators may not have fantastic performance before V8 5.6, but the
difference looks negligible for our purposes.

Closes #5, and also fixes a bug where ordering wasn't even being
respected at all.