Add filters and flags on event objects
Deprecates hook `opts.type`. Adds `opts.filter` which takes ternary
values (true, false, null) for `fake`, `incoming`, `modified`, and
`silenced`, as well as adds properties to the `Buffer` or event object.
The properties have the same names as the flags but with a dollar sign
`$` prepended. The old parameters are still passed for compatibility.

Silenced packets can now be hooked and unsilenced if the hook returns
`true`. Unfortunately this triggers re-serialization and marks it as
modified, so maybe an alternative can be implemented in the future.

Closes #6 and closes #7.